+++ .eslintrc.json
@@ -1,6 +1,6 @@
 {
   "rules": {
-    "max-depth": ["error", { "max": 2 }]
+    "max-lines-per-function": ["error", { "max": 10 }]
   },
   "env": {
     "es6": true,
@@ -10,13 +10,5 @@
     "ecmaVersion": "latest",
     "sourceType": "module"
   },
-  "overrides": [
-    {
-      "files": ["*.test.js", "console.js"],
-      "rules": {
-        "max-lines-per-function": ["off"]
-      }
-    }
-  ],
-  "extends": ["eslint:recommended"]
+  "extends": ["eslint:recommended", "plugin:prettier/recommended"]
 }
